TOT-55 :  [REFACTOR] 유틸성 클래스파일 패키지 분리
